reorganized ve and 3L hydro files
