+++ sph/include/sph/cuda/cuda_density.cu
@@ -3,7 +3,7 @@
 #include "sph.cuh"
 #include "sph/particles_data.hpp"
 #include "cuda_utils.cuh"
-#include "sph/kernel/density_kern.hpp"
+#include "sph/hydro_3L/kernel/density_kern.hpp"
 
 #include "cstone/cuda/findneighbors.cuh"
