+++ sph/include/sph/hydro_3L/density.hpp
@@ -6,7 +6,7 @@
 
 #include "kernel/density_kern.hpp"
 #ifdef USE_CUDA
-#include "cuda/sph.cuh"
+#include "sph/cuda/sph.cuh"
 #endif
 
 namespace sphexa
