+++ sph/test/kernel/momentum_energy.cpp
@@ -33,7 +33,7 @@
 
 #include "gtest/gtest.h"
 
-#include "sph/kernel/momentum_energy_kern.hpp"
+#include "sph/hydro_3L/kernel/momentum_energy_kern.hpp"
 #include "sph/tables.hpp"
 
 using namespace sphexa;
